# swagger-schema.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/otima/webhook/messages": {
      "post": {
        "description": "Receives inbound WhatsApp messages from Otima broker and maps them into the internal message flow.",
        "operationId": "OtimaWebhookController_handleMessages",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Inbound message array payload from Otima",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully"
          }
        },
        "summary": "Handle inbound WhatsApp messages from Otima",
        "tags": [
          "Otima Webhook"
        ]
      }
    },
    "/otima/webhook/ack": {
      "post": {
        "description": "Receives message status updates from Otima broker.",
        "operationId": "OtimaWebhookController_handleAck",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Status update payload from Otima",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status webhook processed successfully"
          }
        },
        "summary": "Handle WhatsApp ACK/status updates from Otima",
        "tags": [
          "Otima Webhook"
        ]
      }
    },
...
    "/messages": {
      "get": {
        "description": "Retrieve messages from PostgreSQL with various filters and pagination options.",
        "operationId": "MessagesController_getMessages",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": false,
            "in": "query",
            "description": "Filter by session ID",
            "schema": {
              "example": "session_123456789",
              "type": "string"
            }
          },
          {
            "name": "senderType",
            "required": false,
            "in": "query",
            "description": "Filter by sender type",
            "schema": {
              "example": "customer",
              "type": "string",
              "enum": [
                "system",
                "bot",
                "customer",
                "user"
              ]
            }
          },
          {
            "name": "recipientType",
            "required": false,
            "in": "query",
            "description": "Filter by recipient type",
            "schema": {
              "example": "user",
              "type": "string",
              "enum": [
                "system",
                "bot",
                "customer",
                "user"
              ]
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "description": "Filter by customer ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter by user ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "name": "platform",
            "required": false,
            "in": "query",
            "description": "Filter by platform",
            "schema": {
              "example": "whatsapp",
              "type": "string",
              "enum": [
                "whatsapp",
                "telegram",
                "instagram",
                "facebook",
                "other"
              ]
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by message type",
            "schema": {
              "example": "text",
              "type": "string",
              "enum": [
                "text",
                "image",
                "video",
                "audio",
                "document",
                "location",
                "contact",
                "sticker",
                "other"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by message status",
            "schema": {
              "example": "sent",
              "type": "string",
              "enum": [
                "pending",
                "sent",
                "delivered",
                "read",
                "failed",
                "deleted"
              ]
            }
          },
          {
            "name": "fromMe",
            "required": false,
            "in": "query",
            "description": "Filter by fromMe flag",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "system",
            "required": false,
            "in": "query",
            "description": "Filter by system flag",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "isGroup",
            "required": false,
            "in": "query",
            "description": "Filter by group flag",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search in message content",
            "schema": {
              "example": "hello",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort by field",
            "schema": {
              "default": "sentAt",
              "example": "sentAt",
              "type": "string",
              "enum": [
                "sentAt",
                "createdAt",
                "updatedAt"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "desc",
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMessagesResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get messages with filtering and pagination",
        "tags": [
          "Messages"
        ]
      }
    },
...
  },
  "info": {
    "title": "CCS Backend API",
    "description": "Multi platform CRM API for customer service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "Typebot-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Typebot API Key (Bearer token)"
      }
    },
    "schemas": {
...
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Message ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "messageId": {
            "type": "string",
            "description": "Platform-specific message identifier",
            "example": "msg_123456789"
          },
          "sessionId": {
            "type": "string",
            "description": "Session identifier",
            "example": "session_123456789"
          },
          "senderType": {
            "type": "string",
            "description": "Sender type",
            "enum": [
              "system",
              "bot",
              "customer",
              "user"
            ],
            "example": "customer"
          },
          "recipientType": {
            "type": "string",
            "description": "Recipient type",
            "enum": [
              "system",
              "bot",
              "customer",
              "user"
            ],
            "example": "user"
          },
          "customerId": {
            "type": "object",
            "description": "Customer ID",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "nullable": true
          },
          "userId": {
            "type": "object",
            "description": "User ID",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "nullable": true
          },
          "fromMe": {
            "type": "boolean",
            "description": "From me flag",
            "example": false
          },
          "system": {
            "type": "boolean",
            "description": "System message flag",
            "example": false
          },
          "isGroup": {
            "type": "boolean",
            "description": "Group message flag",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "Message content",
            "example": "Hello, how can I help you?"
          },
          "media": {
            "type": "string",
            "description": "Media URL",
            "example": "https://example.com/image.jpg"
          },
          "type": {
            "type": "string",
            "description": "Message type",
            "enum": [
              "text",
              "image",
              "video",
              "audio",
              "document",
              "location",
              "contact",
              "sticker",
              "other"
            ],
            "example": "text"
          },
          "platform": {
            "type": "string",
            "description": "Platform",
            "enum": [
              "whatsapp",
              "telegram",
              "instagram",
              "facebook",
              "other"
            ],
            "example": "whatsapp"
          },
          "status": {
            "type": "string",
            "description": "Message status",
            "enum": [
              "pending",
              "sent",
              "delivered",
              "read",
              "failed",
              "deleted"
            ],
            "example": "sent"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata (optional) - may contain platform response data",
            "example": null,
            "nullable": true
          },
          "replyMessageId": {
            "type": "string",
            "description": "Reply message ID",
            "example": "msg_123456789"
          },
          "sentAt": {
            "format": "date-time",
            "type": "string",
            "description": "Sent timestamp",
            "example": "2024-01-01T10:00:00.000Z"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Created timestamp",
            "example": "2024-01-01T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Updated timestamp",
            "example": "2024-01-01T10:00:00.000Z"
          },
          "isText": {
            "type": "boolean",
            "description": "Is text message",
            "example": true
          },
          "isMedia": {
            "type": "boolean",
            "description": "Is media message",
            "example": false
          },
          "isFromCustomer": {
            "type": "boolean",
            "description": "Is from customer",
            "example": true
          },
          "isGroupMessage": {
            "type": "boolean",
            "description": "Is group message",
            "example": false
          },
          "hasReply": {
            "type": "boolean",
            "description": "Has reply",
            "example": false
          },
          "hasMedia": {
            "type": "boolean",
            "description": "Has media",
            "example": false
          }
        },
        "required": [
          "id",
          "messageId",
          "sessionId",
          "senderType",
          "recipientType",
          "customerId",
          "userId",
          "fromMe",
          "system",
          "isGroup",
          "type",
          "platform",
          "status",
          "sentAt",
          "createdAt",
          "updatedAt",
          "isText",
          "isMedia",
          "isFromCustomer",
          "isGroupMessage",
          "hasReply",
          "hasMedia"
        ]
      },
...
      "PaginatedMessagesResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Array of messages",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageResponseDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of messages",
            "example": 100
          },
          "page": {
            "type": "number",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Number of messages per page",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of pages",
            "example": 5
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
...
    }
  }
}